+++ src/lib/index.ts
@@ -1,3 +1,8 @@
+/**
+ * MIT License
+ * Copyright (c) 2023 Yan
+ */
+
 export function isString(n: unknown) {
   return typeof n === "string" || n instanceof String;
 }